+++ clickhouse_connect/driver/httpclient.py
@@ -34,6 +34,7 @@ from clickhouse_connect.driver.transform import NativeTransform
 logger = logging.getLogger(__name__)
 columns_only_re = re.compile(r'LIMIT 0\s*$', re.IGNORECASE)
 ex_header = 'X-ClickHouse-Exception-Code'
+ex_tag_header = 'X-ClickHouse-Exception-Tag'
 
 
 # pylint: disable=too-many-instance-attributes
@@ -299,7 +300,8 @@ class HttpClient(Client):
                                      retries=self.query_retries,
                                      fields=fields,
                                      server_wait=not context.streaming)
-        byte_source = RespBuffCls(ResponseSource(response))  # pylint: disable=not-callable
+        exception_tag = response.headers.get(ex_tag_header)
+        byte_source = RespBuffCls(ResponseSource(response, exception_tag=exception_tag))  # pylint: disable=not-callable
         context.set_response_tz(self._check_tz_change(response.headers.get('X-ClickHouse-Timezone')))
         query_result = self._transform.parse_response(byte_source, context)
         query_result.summary = self._summary(response)